back-end/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
rdb2rdf
python-dotenv
pandas